% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary_mat.R
\name{calc_bern_prob_binary}
\alias{calc_bern_prob_binary}
\title{Calculate a bernoulli probability for assigning e_it = 0 v. 1.}
\usage{
calc_bern_prob_binary(evec, Gvec, beta, betavec, y, theta)
}
\arguments{
\item{evec}{a vector e[-i, t] from the e matrix}

\item{Gvec}{a vector G[-i, i] from the graph matrix G}

\item{beta}{a scalar (on the diagonal) beta[i, i] from beta matrix}

\item{betavec}{a vector betamat[-i, i] from beta matrix}

\item{y}{data matrix}

\item{theta}{sampling parameters, a list}
}
\description{
Calculate a bernoulli probability for assigning e_it = 0 v. 1.
}

