% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{calc_beta_prop}
\alias{calc_beta_prop}
\title{Calculate a proposal for beta matrix's diagonal vector}
\usage{
calc_beta_prop(beta, sd)
}
\arguments{
\item{beta}{a vector of diagonal entries from a beta matrix}

\item{sd}{standard deviation for the distribution of the jump}
}
\description{
Calculate a proposal for beta matrix's diagonal vector
}

