% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary_mat.R
\name{calc_exp_binary}
\alias{calc_exp_binary}
\title{Calculate the exponential factor in updating e}
\usage{
calc_exp_binary(e, evec, Gvec, beta, betavec)
}
\arguments{
\item{e}{a binary scalar}

\item{evec}{a vector e[-i, t] from e matrix}

\item{Gvec}{a vector G[-i, i] from G graph matrix}

\item{beta}{a scalar from the diagonal of beta matrix}

\item{betavec}{a vector betamatrix[-i, i] from beta matrix}
}
\value{
a single number
}
\description{
Calculate the exponential factor in updating e
}

