% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary_mat.R
\name{calc_lik_y}
\alias{calc_lik_y}
\title{Calculate the likelihood given y, after equation 5 of Mitra et al. 2013}
\usage{
calc_lik_y(yvec, evec, theta)
}
\arguments{
\item{theta}{a list of sampling parameters}

\item{y}{a scalar entry from y matrix}

\item{e}{a scalar entry from binary e matrix}
}
\value{
a likelihood, a scalar
}
\description{
Calculate the likelihood given y, after equation 5 of Mitra et al. 2013
}

