% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{calc_logist_prob}
\alias{calc_logist_prob}
\title{Calculate log of unnormalized logistic probability}
\usage{
calc_logist_prob(v, beta)
}
\arguments{
\item{v}{a binary vector}

\item{beta}{a beta matrix of coefficients}
}
\description{
Calculate log of unnormalized logistic probability
}

