% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{calc_RR}
\alias{calc_RR}
\title{Estimate normalizing constant ratio, RR}
\usage{
calc_RR(K = 5000, beta, beta_prop)
}
\arguments{
\item{K}{number of v vectors to sample}

\item{beta}{a beta matrix}

\item{beta_prop}{a proposed beta matrix}
}
\description{
Estimate normalizing constant ratio, RR
}

