% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{calc_sampling_probs}
\alias{calc_sampling_probs}
\title{Calculate logistic probabilities for all binary vectors of a given length}
\usage{
calc_sampling_probs(beta)
}
\arguments{
\item{beta}{a beta matrix of coefficients}
}
\description{
Calculate logistic probabilities for all binary vectors of a given length
}

