% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{sample_binary}
\alias{sample_binary}
\title{Calculate logistic probabilities for all binary vectors of a given length}
\usage{
sample_binary(beta)
}
\arguments{
\item{beta}{a beta matrix of coefficients}
}
\description{
Calculate logistic probabilities for all binary vectors of a given length
}

