% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{sample_columns}
\alias{sample_columns}
\title{Sample columns according to specified probabilities}
\usage{
sample_columns(probs, vecmat, K = 5000)
}
\arguments{
\item{probs}{probabilities vector}

\item{vecmat}{a matrix from which to sample columns}

\item{K}{number of times to sample}
}
\description{
Sample columns according to specified probabilities
}

