% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beta.R
\name{update_beta}
\alias{update_beta}
\title{Update beta}
\usage{
update_beta(beta, emat, s = 0.1, K = 5000)
}
\arguments{
\item{beta}{a beta matrix}

\item{emat}{e binary matrix}

\item{s}{standard deviation for jump}

\item{K}{number of binary vectors to sample}
}
\description{
Update beta
}

