% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary_mat.R
\name{update_binary}
\alias{update_binary}
\title{Update a single element of the binary indicator matrix e}
\usage{
update_binary(evec, Gvec, beta, betavec, y, theta)
}
\arguments{
\item{evec}{vector e_-it in Mitra 2013 notation}

\item{Gvec}{vector G_-it in Mitra 2013 notation}

\item{beta}{a scalar, entry [i, i] of beta matrix}

\item{betavec}{a vector from beta matrix, beta_matrix[i, -i]}

\item{y}{scalar entry from data matrix}

\item{theta}{sampling distribution parameters, a list}
}
\value{
a binary scalar (0 or 1)
}
\description{
Update a single element of the binary indicator matrix e
}

