% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_G.R
\name{update_G}
\alias{update_G}
\title{Update graph, G}
\usage{
update_G(G, betamat, emat)
}
\arguments{
\item{G}{a graph matrix}

\item{betamat}{a beta matrix}

\item{emat}{an e matrix of binary indicators}
}
\description{
Update graph, G
}

