% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_G.R
\name{update_graph}
\alias{update_graph}
\title{Update graph, G (and beta matrix)}
\usage{
update_graph(graph, betamat, emat)
}
\arguments{
\item{betamat}{a beta matrix}

\item{emat}{an e matrix of binary indicators}

\item{G}{a graph matrix}
}
\description{
Update graph, G (and beta matrix)
}

