% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lambda.R
\name{update_lambda}
\alias{update_lambda}
\title{Update the lambda vector}
\usage{
update_lambda(lambdavec, ymat, zmat, a = 0.001, b = 0.001)
}
\arguments{
\item{lambdavec}{a lambda vector of Poisson means}

\item{ymat}{a y data matrix}

\item{zmat}{a trinary z indicator matrix}

\item{a}{gamma hyperprior}

\item{b}{gamma hyperprior}
}
\description{
Update the lambda vector
}

