% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{update_mu}
\alias{update_mu}
\title{Update a mu vector (mu1 or mu2)}
\usage{
update_mu(muvec, sigmavec, ymat, zmat, indicator, mu0, t0)
}
\arguments{
\item{muvec}{mu vector to be updated}

\item{sigmavec}{sd vector (goes with mu vector)}

\item{ymat}{data matrix}

\item{zmat}{a trinary indicator matrix}

\item{indicator}{(0 or 1) indicator to denote which mu vector (1 or 2) is being updated}

\item{mu0}{hyperparameter, mean of normal hyperprior}

\item{t0}{hyperparameter, precision of normal hyperprior}
}
\description{
Update a mu vector (mu1 or mu2)
}

