% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sigma.R
\name{update_sigma}
\alias{update_sigma}
\title{Update a sigma vector}
\usage{
update_sigma(sigmavec, ymat, muvec, zmat, indicator, a0 = 0.001, b0 = 0.001)
}
\arguments{
\item{sigmavec}{an inputted sigmavec}

\item{ymat}{the data matrix}

\item{muvec}{the 'corresponding' means vector}

\item{zmat}{matrix of trinary indicators}

\item{indicator}{0 or 1 to denote which pair of means and sd's are of interest}

\item{a0}{hyperparameter, shape parameter of hyperprior}

\item{b0}{hyperparameter, rate parameter of hyperprior}
}
\description{
Update a sigma vector
}

