% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trinary_mat.R
\name{update_trinary_mat}
\alias{update_trinary_mat}
\title{Wrapper to update trinary matrix}
\usage{
update_trinary_mat(emat, ymat, mu1vec, mu2vec, sd1vec, sd2vec)
}
\arguments{
\item{emat}{a binary matrix}

\item{ymat}{a data matrix}

\item{mu1vec}{a mu vector}

\item{mu2vec}{a mu vector}

\item{sd1vec}{a sd vector}

\item{sd2vec}{a sd vector}
}
\description{
Wrapper to update trinary matrix
}

