% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trinary_mat.R
\name{update_trinary_scalar}
\alias{update_trinary_scalar}
\title{Update trinary scalar}
\usage{
update_trinary_scalar(e, y, pp, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{e}{a scalar, with value 0 or 1}

\item{y}{a scalar}

\item{pp}{a scalar probability}

\item{mu1}{a scalar mean}

\item{mu2}{a scalar mean}

\item{sigma1}{a scalar sd}

\item{sigma2}{a scalar sd}
}
\description{
Update trinary scalar
}

