% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trinary_mat.R
\name{update_trinary_vector}
\alias{update_trinary_vector}
\title{Wrapper to update a column of z matrix}
\usage{
update_trinary_vector(evec, yvec, mu1vec, mu2vec, sigma1vec, sigma2vec)
}
\arguments{
\item{evec}{a column from the e matrix}

\item{yvec}{a column from the y matrix}

\item{mu1vec}{a mu vector}

\item{mu2vec}{a mu vector}

\item{sd1vec}{a sd vector}

\item{sd2vec}{a sd vector}
}
\description{
Wrapper to update a column of z matrix
}

