% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_binary_mat.R
\name{update_binary_mat}
\alias{update_binary_mat}
\title{Wrapper function to update binary matrix}
\usage{
update_binary_mat(emat, Gmat, betamat, y, theta)
}
\arguments{
\item{emat}{a binary matrix}

\item{Gmat}{a graph matrix}

\item{betamat}{a beta matrix}

\item{y}{data matrix}

\item{theta}{sampling parameters, as a list}
}
\value{
binary matrix of same dimensions as emat, ie, an updated e matrix
}
\description{
Wrapper function to update binary matrix
}

